Revert "Let H2 run queries again."
This reverts commit 2c7e888a09e7c2124fcad594648219b1b8622f8b.

+++ src/main/java/me/moodcat/database/controllers/SongDAO.java
@@ -104,7 +104,7 @@ public class SongDAO extends AbstractDAO<Song> {
      * @return List of songs
      */
     @Transactional
-    public List<Song> findForDistance(final VAVector vector, final int limit) {
+    public List<Song> findForDistance(final VAVector vector, final long limit) {
         return query().from(song)
                 .orderBy(song.valenceArousal.location.distance(vector.getLocation()).asc())
                 .limit(limit)
